requirements.txt
altgraph==0.17.4
cx-Oracle==8.3.0
et-xmlfile==1.1.0
greenlet==3.0.3
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1